Initial pass at rethinking the toolbar
Imports 32678.5.diff from the core trac ticket. Needs some CSS debugging
in terms of the styles, especially around the "admin" menu in the site
title dropdown on the frontend, but contains most of the concepts and
designed discussed on the ticket. After a bit more cleanup, should be
ready for testing and iterations.